+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2026-01-27]
+
+### Refactored
+- Split HarvestApiService into focused services (HarvestOAuthService, HarvestTokenManager, HarvestApiClient, HarvestConfiguration)
+- Moved Harvest models to Models/Application/Harvest folder
+- Converted db-access command to Cursor skill
+
 ## [2026-01-26]
 
 ### Added